AP-9669: Revert commit 10fde27514252933bd984d45ec7c18c941e97db1 as it is not backwards compatible
AP-9669 (DL Keras: Node Generation Spinner Widget Bugs)

+++ org.knime.dl.keras/src/org/knime/dl/keras/core/config/constraint/DLKerasMaxNormConstraint.java
@@ -59,7 +59,7 @@ import org.knime.dl.python.util.DLPythonUtils;
 public final class DLKerasMaxNormConstraint extends DLKerasAbstractAxisConstraint {
 
     @Parameter(label = "Maximum norm", min = "0.0000001", stepSize = "0.0000001")
-    private double m_maxValue = 2.0;
+    private float m_maxValue = 2.0f;
 
     /**
      */
