+++ org.knime.dl.keras/src/org/knime/dl/keras/core/config/regularizer/DLKerasL1Regularizer.java
@@ -60,7 +60,7 @@ import org.knime.dl.python.util.DLPythonUtils;
 public final class DLKerasL1Regularizer extends DLKerasAbstractConfigObject implements DLKerasRegularizer {
 
     @Parameter(label = "L1 regularization factor", min = "0.0000001", stepSize = "0.0000001")
-    private double m_l1 = 0.0000001f;
+    private float m_l1 = 0.0000001f;
 
     /**
      */
